indicatiotion that there is somthing in queue
